Move logger to core
