+++ src/core/sdk.js
@@ -8,7 +8,7 @@ import Account from './Account';
 import MPEContract from './MPEContract';
 import { find } from 'lodash';
 import DefaultPaymentChannelManagementStrategy from './payment_channel_management_strategies';
-import logger from '../utils/logger';
+import logger from './utils/logger';
 
 const DEFAULT_CONFIG = {
   defaultGasLimit: 210000,
