+++ src/nodejs/index.js
@@ -1,3 +1,4 @@
+import './utils/logger';
 import DefaultPaymentChannelManagementStrategy from '../core/payment_channel_management_strategies';
 import NodeSdk from './NodeSdk';
